requirements.txt
requests
pandas
yfinance
streamlit
python-dotenv
schedule
python-telegram-bot
